LayoutTests/traversal/moz-bug559526-expected.txt

CONSOLE MESSAGE: line 49: Should have thrown an exception: null
CONSOLE MESSAGE: line 49: Should have thrown an exception: null
CONSOLE MESSAGE: line 49: Should have thrown an exception: null
CONSOLE MESSAGE: line 49: Should have thrown an exception: null
CONSOLE MESSAGE: line 94: Should have tests 4 filter calls: 4
Mozilla Bug 559526
